Update internal race value to 64 bit

+++ card.h
@@ -45,7 +45,7 @@ struct card_state {
 	uint32_t lscale{};
 	uint32_t rscale{};
 	uint32_t attribute{};
-	uint32_t race{};
+	uint64_t race{};
 	int32_t attack{};
 	int32_t defense{};
 	int32_t base_attack{};
@@ -133,7 +133,7 @@ public:
 	uint32_t spsummon_code{};
 	uint16_t spsummon_counter[2]{};
 	uint16_t spsummon_counter_rst[2]{};
-	std::map<uint32_t, uint32_t> assume;
+	std::map<uint32_t, uint64_t> assume;
 	card* equiping_target{};
 	card* pre_equip_target{};
 	card* overlay_target{};
@@ -189,7 +189,7 @@ public:
 	uint32_t get_ritual_level(card* pcard);
 	uint32_t check_xyz_level(card* pcard, uint32_t lv);
 	uint32_t get_attribute(card* scard = 0, uint64_t sumtype = 0, uint8_t playerid = 2);
-	uint32_t get_race(card* scard = 0, uint64_t sumtype = 0, uint8_t playerid = 2);
+	uint64_t get_race(card* scard = 0, uint64_t sumtype = 0, uint8_t playerid = 2);
 	uint32_t get_lscale();
 	uint32_t get_rscale();
 	uint32_t get_link_marker();
